refactored hl-cli into multiple files
